Consolidate user/group privilege tests into executor_test.go
• Merged usergroup_test.go content into executor_test.go following Go naming
  conventions where test files match their source files (executor.go)
• Added //go:build test tag to enable access to PrepareCommand helper function
• Deleted standalone usergroup_test.go file which had no corresponding
  usergroup.go source file
• Consolidated 12 test functions covering user/group privilege execution
  scenarios including success paths, error handling, and audit logging
• All tests continue to pass with improved file organization structure
